[Functions] Move project output patching to backend, no need for this in frontend
