+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/testData/org/cases/documentmodel/completion/csharp/defaultWorker/AzureFunctionsTimerTriggerCompletionTest/testPopup_CompletionList_FullListOnBasicCompletion/source/Function.cs
