+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/testData/org/cases/documentmodel/completion/csharp/isolatedWorker/AzureFunctionsTimerTriggerCompletionTest/testPopup_CompletionList_FullListOnBasicCompletion/gold/Function.gold
@@ -1,0 +1,17 @@
+using System;
+using System.Threading.Tasks;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Extensions.Logging;
+
+namespace FunctionAppIsolated
+{
+    public static class Function
+    {
+        [Function("TimerTrigger")]
+        public static async Task RunAsync([TimerTrigger("*/15 * * * * *<caret>")] TimerInfo myTimer, ILogger log)
+        {
+            log.LogInformation($"C# Timer trigger function executed at: {DateTime.UtcNow}");
+            
+        }
+    }
+}
